Tests/LibWeb/Text/input/wpt-import/encoding-detection/is-windows-1252.tentative.html

<!doctype html>
<title>is windows-1252</title>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<p>Ţetta er kóđunarpróf á staf. Fyrir sum tungumál sem nota latneska stafi ţurfum viđ meira inntak til ađ taka ákvörđunina.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1252", 'Expected windows-1252');
    }, "Check detection result");
      done();
};
</script>
